refactor swaps tests
